added functionality for adding materials to assembly

+++ model/src/main/java/com/promin_ism/model/Assembly.java
@@ -200,7 +200,7 @@ public class Assembly {
         if (materials.containsKey(material)){
             return false;
         }
-        if(quantity == null || quantity.longValue() <= 0 ){
+        if(quantity == null || quantity.floatValue() <= 0){
             return false;
         }
         else{
